### 2025-09-19 #9027 @editorialbot
[PRE REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R
**Submitting author:** @ (<a href="http://orcid.org/0000-0001-7719-2619">Olatunde David Akanbi</a>)
**Repository:** https://github.com/cwru-sdle/geospatialsuite
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
