+++ paper/paper.bib
@@ -97,7 +97,7 @@
   pages={1053},
   year={2025},
   publisher={Nature Publishing Group UK London},
-  doi={10.1038/s41598-024-84771-4}
+  doi={10.1038/s41598-024-85050-3}
 }
 
 @inproceedings{akanbi2024integrating,
@@ -119,7 +119,7 @@
   pages={9},
   year={2024},
   publisher={Springer},
-  doi={10.1007/s41651-024-00170-3}
+  doi={10.1007/s41651-023-00164-y}
 }
 
 @mastersthesis{akanbi2024leveraging,
