### 2025-09-28 daf002b
updated the vegetation indices
