+++ README.md
@@ -13,7 +13,8 @@
 
 [![CRAN
 status](https://www.r-pkg.org/badges/version/geospatialsuite)](https://CRAN.R-project.org/package=geospatialsuite)
-[![Downloads](https://cranlogs.r-pkg.org/badges/geospatialsuite)](https://cran.r-project.org/package=geospatialsuite)
+[![Total
+Downloads](https://cranlogs.r-pkg.org/badges/grand-total/geospatialsuite)](https://cran.r-project.org/package=geospatialsuite)
 [![License:
 MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 <!-- badges: end -->
@@ -379,68 +380,73 @@ result <- universal_spatial_join(
 
 <summary>
 
-<strong>Click to see all 60+ vegetation indices</strong>
+<strong>Click to see all 60+ vegetation indices (full table)</strong>
 </summary>
 
-### Basic Vegetation Indices
-
-- **NDVI** - Normalized Difference Vegetation Index (Rouse et al. 1974)
-- **EVI** - Enhanced Vegetation Index (Huete et al. 2002)
-- **SAVI** - Soil Adjusted Vegetation Index (Huete 1988)
-- **MSAVI** - Modified Soil Adjusted Vegetation Index (Qi et al. 1994)
-- **OSAVI** - Optimized Soil Adjusted Vegetation Index (Rondeaux et
-  al. 1996)
-- **GNDVI** - Green Normalized Difference Vegetation Index
-
-### Atmospheric Correction Indices
-
-- **ARVI** - Atmospherically Resistant Vegetation Index (Kaufman & Tanré
-  1992)
-- **RDVI** - Renormalized Difference Vegetation Index
-- **WDVI** - Weighted Difference Vegetation Index
-- **PVI** - Perpendicular Vegetation Index
-- **VARI** - Visible Atmospherically Resistant Index
-
-### Red-Edge Indices (Sentinel-2/Hyperspectral)
-
-- **NDRE** - Normalized Difference Red Edge
-- **MTCI** - MERIS Terrestrial Chlorophyll Index
-- **IRECI** - Inverted Red-Edge Chlorophyll Index
-- **S2REP** - Sentinel-2 Red-Edge Position
-- **CIred-edge** - Chlorophyll Index Red-Edge
-
-### Stress Detection and Pigment Indices
-
-- **PRI** - Photochemical Reflectance Index
-- **SIPI** - Structure Insensitive Pigment Index
-- **PSRI** - Plant Senescence Reflectance Index
-- **CCI** - Canopy Chlorophyll Index
-- **ARI** - Anthocyanin Reflectance Index
-
-### Water and Moisture Indices
-
-- **NDWI** - Normalized Difference Water Index (McFeeters 1996)
-- **MNDWI** - Modified Normalized Difference Water Index (Xu 2006)
-- **NDMI** - Normalized Difference Moisture Index (Gao 1996)
-- **MSI** - Moisture Stress Index
-- **WI** - Water Index
-- **LSWI** - Land Surface Water Index
-
-### Specialized Agricultural Indices
-
-- **GCCI** - Green Crop Coverage Index
-- **NGRDI** - Normalized Green-Red Difference Index  
-- **GLI** - Green Leaf Index
-- **VDVI** - Visible-band Difference Vegetation Index
-- **TGI** - Triangular Greenness Index
-
-### Advanced Biophysical Indices
-
-- **LAI** - Leaf Area Index estimators
-- **FAPAR** - Fraction of Absorbed PAR
-- **MTVI1/MTVI2** - Modified Triangular Vegetation Index
[…]
