+++ README.Rmd
@@ -27,7 +27,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: start -->
 [![CRAN status](https://www.r-pkg.org/badges/version/geospatialsuite)](https://CRAN.R-project.org/package=geospatialsuite)
-[![Downloads](https://cranlogs.r-pkg.org/badges/geospatialsuite)](https://cran.r-project.org/package=geospatialsuite)
+[![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/geospatialsuite)](https://cran.r-project.org/package=geospatialsuite)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 <!-- badges: end -->
 
@@ -345,59 +345,73 @@ result <- universal_spatial_join(
 
 ## 📊 Supported Vegetation Indices
 
-<details>
-<summary><strong>Click to see all 60+ vegetation indices</strong></summary>
+<details> <summary><strong>Click to see all 60+ vegetation indices (full table)</strong></summary>
 
-### Basic Vegetation Indices
-- **NDVI** - Normalized Difference Vegetation Index (Rouse et al. 1974)
-- **EVI** - Enhanced Vegetation Index (Huete et al. 2002)
-- **SAVI** - Soil Adjusted Vegetation Index (Huete 1988)
-- **MSAVI** - Modified Soil Adjusted Vegetation Index (Qi et al. 1994)
-- **OSAVI** - Optimized Soil Adjusted Vegetation Index (Rondeaux et al. 1996)
-- **GNDVI** - Green Normalized Difference Vegetation Index
+| #  | Index      | Category    | Application | Required Bands   | Description                                    |
+| -- | ---------- | ----------- | ----------- | ---------------- | ---------------------------------------------- |
+| 1  | **NDVI**   | basic       | general     | Red, NIR         | Normalized Difference Vegetation Index         |
+| 2  | **SAVI**   | basic       | soil        | Red, NIR         | Soil Adjusted Vegetation Index                 |
+| 3  | **MSAVI**  | basic       | soil        | Red, NIR         | Modified Soil Adjusted Vegetation Index        |
+| 4  | **OSAVI**  | basic       | soil        | Red, NIR         | Optimized Soil Adjusted Vegetation Index       |
+| 5  | **EVI**    | basic       | general     | Red, NIR, Blue   | Enhanced Vegetation Index                      |
+| 6  | **EVI2**   | basic       | general     | Red, NIR         | Two-band Enhanced Vegetation Index             |
+| 7  | **DVI**    | basic       | biomass     | Red, NIR         | Difference Vegetation Index                    |
+| 8  | **RVI**    | basic       | general     | Red, NIR         | Ratio Vegetation Index                         |
+| 9  | **GNDVI**  | basic       | chlorophyll | Green, NIR       | Green NDVI                                     |
+| 10 | **WDVI**   | basic       | soil        | Red, NIR         | Weighted Difference Vegetation Index           |
+| 11 | **ARVI**   | enhanced    | atmospheric | Red, NIR, Blue   | Atmospherically Resistant Vegetation Index     |
+| 12 | **RDVI**   | enhanced    | general     | Red, NIR         | Renormalized Difference Vegetation Index       |
+| 13 | **PVI**    | enhanced    | general     | Red, NIR         | Perpendicular Vegetation Index                 |
+| 14 | **IPVI**   | enhanced    | general     | Red, NIR         | Infrared Percentage Vegetation Index           |
+| 15 | **TNDVI**  | enhanced    | general     | Red, NIR         | Transformed NDVI                               |
[…]
