### 2025-11-06 bb7391f
new version release with auto-geocoding and extensive documentation
