+++ R/15-helpers.R
@@ -310,7 +310,7 @@ add_lines_to_leaflet <- function(map_obj, spatial_data, fill_variable, popup_var
   return(map_obj)
 }
 
-#' Run enhanced terrain analysis workflow - IMPLEMENTED
+#' Run enhanced terrain analysis workflow
 #' @keywords internal
 run_enhanced_terrain_analysis_workflow <- function(config, output_folder) {
   message("Running enhanced terrain analysis workflow...")
@@ -364,7 +364,7 @@ run_enhanced_terrain_analysis_workflow <- function(config, output_folder) {
   return(terrain_results)
 }
 
-#' Run enhanced temporal workflow - IMPLEMENTED
+#' Run enhanced temporal workflow
 #' @keywords internal
 run_enhanced_temporal_workflow <- function(config, output_folder) {
   message("Running enhanced temporal workflow...")
@@ -416,7 +416,7 @@ run_enhanced_temporal_workflow <- function(config, output_folder) {
   return(results)
 }
 
-#' Run enhanced mosaic workflow - IMPLEMENTED
+#' Run enhanced mosaic workflow
 #' @keywords internal
 run_enhanced_mosaic_workflow <- function(config, output_folder) {
   message("Running enhanced mosaic workflow...")
@@ -454,7 +454,7 @@ run_enhanced_mosaic_workflow <- function(config, output_folder) {
   return(results)
 }
 
-#' Run multi-dataset workflow - IMPLEMENTED
+#' Run multi-dataset workflow
 #' @keywords internal
 run_multi_dataset_workflow <- function(config, output_folder) {
   message("Running multi-dataset workflow...")
@@ -642,12 +642,477 @@ load_vector_data_safe <- function(file_path) {
   }
 }
 
+
+
+
+# ==================== GEOGRAPHIC ENTITY DETECTION & GEOCODING ==================== #
+
+#' Normalize column names for robust matching
+#' @keywords internal
+normalize_column_name <- function(x) {
+  tolower(gsub("[_\\s-]", "", x))
+}
+
+#' Detect and geocode geographic entity columns
+#' @param data_input Data frame with potential geographic identifiers
+#' @return List with detected column and type, or NULL if none found
+#' @keywords internal
+detect_geographic_entities <- function(data_input) {
+
+  # Normalize column names: lowercase, remove spaces, underscores, hyphens
+  column_names_normalized <- normalize_column_name(names(data_input))
+  column_names_original <- names(data_input)
+
+  # Define patterns for different geographic entities (all normalized)
+  patterns <- list(
+    state = c("state", "statename", "st", "stateabbr", "statecode", "stusps", "stateab"),
+    county = c("county", "countyname", "cnty", "cty", "countyfips"),
+    fips = c("fips", "fipscode", "geoid", "geoidfips", "geofips", "geocode"),
+    huc = c("huc", "huc2", "huc4", "huc6", "huc8", "huc10", "huc12",
+            "huc14", "watershed", "watershedid", "hydrologicunitcode",
+            "hydrologicunit", "basinid", "subbasin"),
+    zipcode = c("zip", "zipcode", "postal", "postalcode", "postcode"),
+    city = c("city", "cityname", "municipality", "town", "townname")
+  )
+
+  # Check each pattern type
+  for (entity_type in names(patterns)) {
+    pattern_list <- patterns[[entity_type]]
+
+    # Check each specific pattern
+    for (pattern in pattern_list) {
+      # Find exact matches first
+      exact_matches <- which(column_names_normalized == pattern)
+
+      if (length(exact_matches) > 0) {
+        return(list(
+          column = column_names_original[exact_matches[1]],
[…]
