+++ cran-comments.md
@@ -2,4 +2,15 @@
 
 0 errors | 0 warnings | 1 note
 
-* This is a new release.
+* ### Documentation
+* Complete formulas for all 62 vegetation indices (previously only 28/62)
+* Added comprehensive band naming guide for Landsat, Sentinel-2, and MODIS
+* Enhanced output interpretation documentation for analyze_crop_vegetation()
+* Added literature references for all indices
+* Clarified composite yield index calculation and interpretation
+* Added auto-geocoding for data without coordinates
+### Bug Fixes
+* Fixed 34 vegetation indices that showed "Formula not available"
+
+### Citation
+* Added JOSS paper and methodology citations

[…]
