### 2025-11-06 0a14782
include contribution statements
