+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,21 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: '[FEATURE] '
+labels: enhancement
+---
+
+**Is your feature request related to a problem?**
+A clear description of what the problem is.
+
+**Describe the solution you'd like**
+A clear description of what you want to happen.
+
+**Describe alternatives you've considered**
+Alternative solutions or features you've considered.
+
+**Use case**
+How would this feature be used? Who would benefit?
+
+**Additional context**
+Add any other context or examples about the feature request here.
