+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,34 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: '[BUG] '
+labels: bug
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+```r
+# Your reproducible code here
+```
+
+**Expected behavior**
+What you expected to happen.
+
+**Actual behavior**
+What actually happened.
+
+**System Information:**
+- OS: [e.g., Windows 10, macOS 13, Ubuntu 22.04]
+- R version: [e.g., 4.3.0]
+- GeoSpatialSuite version: [e.g., 0.1.0]
+- Relevant package versions:
+```r
+  packageVersion("terra")
+  packageVersion("sf")
+```
+
+**Additional context**
+Add any other context about the problem here.
