### 2025-12-02 #9536 @editorialbot
[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R
**Submitting author:** @Exelegch (<a href="http://orcid.org/0000-0001-7719-2619">Olatunde David Akanbi</a>)
**Repository:** https://github.com/cwru-sdle/geospatialsuite
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
