### 2025-12-02 #9536 @Exelegch
@benyamindsmith

> @Exelgch
>
> I just read your paper draft. I have the following preliminary questions:
>
> 1. The repository that you are submitting to review is a fork from another repository. Whats the reason for this?
> 2. Despite the package being quite big, there is very little commit history associated with it. Can you please tell me more about the development process?

Thank you for the review. I have responded to this (initially raised by the editor)
1. Our group develops the software privately in a Bitbucket repository for day-to-day work.
2. We maintain that internal history there and periodically release stable, reviewed versions to GitHub for public distribution.
3. I initially submitted the paper via the cwru-sdle repository (JOSS said I can't use organization repository)
4. The Exelegch/geospatialsuite repository is therefore a fork of our lab’s public GitHub repo so that I can act as the submitting author, but the actual development history resides in our private Bitbucket repository.
5. The short commit history on GitHub reflects only the stable release that we prepared for JOSS.
