### 2025-12-02 #9536 @benyamindsmith
@Exelegch

Thanks for letting me know.

In terms of the commit history- I do need to verify your contributions (as well as the other authors) though. i.e.

 >  - [ ] Contribution and authorship: Has the submitting author (@Exelegch) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@kdorheim let me know if I can check off this point based on the information given or if some more evidence is required.
