### 2025-12-25 #9536 @benyamindsmith
@Exelegch I am working on reviewing your R package but I am unable to open any issues on the repository.  Please enable this and I will open up issues relevant for the review.
