### 2026-01-01 f0b27db
Add pkgdown site, added testing data,  and update DESCRIPTION
