+++ paper/paper.md
@@ -1,6 +1,5 @@
 ---
-title: 'GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration
-  Toolkit for R'
+title: 'geospatialsuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R'
 tags:
 - R
 - geospatial analysis
@@ -12,11 +11,11 @@ tags:
 - sf
 - agricultural research
 - environmental monitoring
-date: "17 September 2025"
+date: "28 December 2025"
 output:
+  pdf_document: default
   html_document:
     df_print: paged
-  pdf_document: default
 bibliography: paper.bib
 authors:
 - name: Olatunde D. Akanbi
@@ -49,9 +48,9 @@ affiliations:
 
 # Summary
 
-GeoSpatialSuite is a comprehensive R package for geospatiotemporal analysis that provides researchers with a unified toolkit for multimodal spatial data integration. The package features over 60 vegetation indices with automatic satellite band detection, universal spatial join operations that work with any raster-vector combination, robust raster visualization with built-in error handling, and rapid mapping capabilities through the `quick_map()` function. Designed for reliability and universal applicability, GeoSpatialSuite addresses critical needs in agricultural research, environmental monitoring, and remote sensing applications while maintaining computational efficiency for large datasets.
+Geospatialsuite is a comprehensive R package for geospatiotemporal analysis that provides researchers with a unified toolkit for multimodal spatial data integration. The package features over 60 vegetation indices with automatic satellite band detection, universal spatial join operations that work with any raster-vector combination, robust raster visualization with built-in error handling, and rapid mapping capabilities through the `quick_map()` function. Designed for reliability and universal applicability, geospatialsuite addresses critical needs in agricultural research, environmental monitoring, and remote sensing applications while maintaining computational efficiency for large datasets.
 
-The package seamlessly integrates with the modern R spatial ecosystem through `terra` [@hijmans2022terra] and `sf` [@pebesma2018simple], providing enhanced NDVI calculations with quality filtering, comprehensive water quality analysis using multiple indices (NDWI [@mcfeeters1996normalized], MNDWI [@xu2006modification], NDMI), crop data layer (CDL) analysis [@usda2024cdl], spatial interpolation techniques [@cressie1993statistics], and terrain analysis. Unlike existing solutions that are often limited by complex dependencies or geographic scope, GeoSpatialSuite emphasizes universal functionality—working seamlessly across different regions, satellite platforms (Landsat, Sentinel-2, MODIS), and data types with comprehensive error handling throughout. The package has enabled research in agricultural systems monitoring [@akanbi2024multiscale] and multimodal data integration [@akanbi2024integrating].
+The package seamlessly integrates with the modern R spatial ecosystem through `terra` [@hijmans2022terra] and `sf` [@pebesma2018simple], providing enhanced NDVI calculations with quality filtering, comprehensive water quality analysis using multiple indices (NDWI [@mcfeeters1996normalized], MNDWI [@xu2006modification], NDMI), crop data layer (CDL) analysis [@usda2024cdl], spatial interpolation techniques [@cressie1993statistics], and terrain analysis. Unlike existing solutions that are often limited by complex dependencies or geographic scope, geospatialsuite emphasizes universal functionality—working seamlessly across different regions, satellite platforms (Landsat, Sentinel-2, MODIS), and data types with comprehensive error handling throughout. The package has enabled research in agricultural systems monitoring [@akanbi2024multiscale] and multimodal data integration [@akanbi2024integrating].
 
 # Statement of Need
 
@@ -59,7 +58,7 @@ Geospatial analysis in environmental and agricultural research requires reliable
 
[…]
