+++ R/13-water-quality.R
@@ -1017,7 +1017,7 @@ generate_statistics_report <- function(statistics, threshold_results, spatial_an
   report_lines <- c(report_lines,
                     "",
                     strrep("=", 50),
-                    "Report generated by GeoSpatialSuite water quality analysis"
+                    "Report generated by geospatialsuite water quality analysis"
   )
 
   return(report_lines)
