### 2026-01-01 e7afb48
Add GitHub Actions for automatic pkgdown deployment
