+++ reference/add_points_to_leaflet.md
@@ -1,0 +1,16 @@
+# Add points to leaflet -
+
+Add points to leaflet -
+
+## Usage
+
+``` r
+add_points_to_leaflet(
+  map_obj,
+  spatial_data,
+  fill_variable,
+  popup_vars,
+  color_scheme,
+  cluster_points
+)
+```
