+++ reference/index.md
@@ -1,0 +1,486 @@
+# Package index
+
+## Package Documentation
+
+Overview of the geospatialsuite package
+
+- [`geospatialsuite`](https://cwru-sdle.github.io/geospatialsuite/reference/geospatialsuite-package.md)
+  [`geospatialsuite-package`](https://cwru-sdle.github.io/geospatialsuite/reference/geospatialsuite-package.md)
+  : geospatialsuite: Comprehensive Geospatiotemporal Analysis and
+  Multimodal Integration Toolkit
+
+## Core Vegetation Indices
+
+Calculate vegetation indices from spectral bands
+
+- [`calculate_vegetation_index()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_vegetation_index.md)
+  : Calculate comprehensive vegetation indices
+- [`calculate_multiple_indices()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_multiple_indices.md)
+  : Calculate multiple vegetation indices at once
+- [`calculate_ndvi_enhanced()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_ndvi_enhanced.md)
+  : Calculate NDVI with time series options
+- [`list_vegetation_indices()`](https://cwru-sdle.github.io/geospatialsuite/reference/list_vegetation_indices.md)
+  : Get comprehensive list of available vegetation indices
+- [`get_available_indices()`](https://cwru-sdle.github.io/geospatialsuite/reference/get_available_indices.md)
+  : Get available indices
+- [`calculate_vegetation_statistics()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_vegetation_statistics.md)
+  : Calculate vegetation statistics
+- [`calculate_comprehensive_vegetation_stats()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_comprehensive_vegetation_stats.md)
+  : Calculate comprehensive vegetation statistics -
+
+## Sample Data Access
+
+Built-in sample datasets for examples and testing
+
+- [`download_sample_data()`](https://cwru-sdle.github.io/geospatialsuite/reference/download_sample_data.md)
+  : Download or access sample datasets for geospatialsuite
+- [`list_sample_datasets()`](https://cwru-sdle.github.io/geospatialsuite/reference/list_sample_datasets.md)
+  : List available sample datasets
+- [`get_sample_data()`](https://cwru-sdle.github.io/geospatialsuite/reference/get_sample_data.md)
+  : Get path to sample data file
+- [`load_sample_data()`](https://cwru-sdle.github.io/geospatialsuite/reference/load_sample_data.md)
+  : Load sample data directly
+
+## Water Quality & Indices
+
+Water indices and quality assessment
+
+- [`calculate_water_index()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_water_index.md)
+  : Calculate water indices including both NDWI variants
+- [`calculate_multiple_water_indices()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_multiple_water_indices.md)
+  : Calculate multiple water indices at once
+- [`analyze_water_quality_comprehensive()`](https://cwru-sdle.github.io/geospatialsuite/reference/analyze_water_quality_comprehensive.md)
+  : Analyze water quality comprehensively with flexible data handling
+- [`calculate_water_quality_statistics()`](https://cwru-sdle.github.io/geospatialsuite/reference/calculate_water_quality_statistics.md)
+  : Calculate comprehensive water quality statistics
[…]
