+++ reference/get_reliable_colors.md
@@ -1,0 +1,9 @@
+# Get reliable colors for terra plotting
+
+Get reliable colors for terra plotting
+
+## Usage
+
+``` r
+get_reliable_colors(color_scheme, n = 100)
+```
