+++ reference/classify_spatial_data.md
@@ -1,0 +1,23 @@
+# Classify spatial data type and load data
+
+Classify spatial data type and load data
+
+## Usage
+
+``` r
+classify_spatial_data(data_input, verbose = FALSE)
+```
+
+## Arguments
+
+- data_input:
+
+  Input data (file path, directory, or spatial object)
+
+- verbose:
+
+  Print progress messages
+
+## Value
+
+List with data, type, class, and path information
