+++ reference/validate_vegetation_analysis.md
@@ -1,0 +1,13 @@
+# Validate vegetation analysis
+
+Validate vegetation analysis
+
+## Usage
+
+``` r
+validate_vegetation_analysis(
+  vegetation_indices,
+  reference_data,
+  verbose = FALSE
+)
+```
