+++ reference/analyze_growth_stage.md
@@ -1,0 +1,9 @@
+# Analyze growth stage
+
+Analyze growth stage
+
+## Usage
+
+``` r
+analyze_growth_stage(vegetation_indices, crop_type, growth_stage, verbose)
+```

[…]
