### 2026-01-01 91c298b
Add GitHub Actions for automatic pkgdown deployment2
