+++ docs/articles/index.md
@@ -3,18 +3,18 @@
 ### All vignettes
 
 - [Agricultural Applications with
-  geospatialsuite](https://cwru-sdle.github.io/geospatialsuite/articles/agricultural-analysis.md):
+  geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/agricultural-analysis.md):
 - [Using analyze_crop_vegetation() in
-  geospatialsuite](https://cwru-sdle.github.io/geospatialsuite/articles/analyze-crop-vegetation.md):
+  geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/analyze-crop-vegetation.md):
 - [Getting Started with
-  geospatialsuite](https://cwru-sdle.github.io/geospatialsuite/articles/getting-started.md):
+  geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/getting-started.md):
 - [Universal Spatial Mapping with
-  geospatialsuite](https://cwru-sdle.github.io/geospatialsuite/articles/spatial-analysis.md):
+  geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/spatial-analysis.md):
 - [Spatial Data
-  Integration](https://cwru-sdle.github.io/geospatialsuite/articles/spatial-integration.md):
+  Integration](https://exelegch.github.io/geospatialsuite-docs/articles/spatial-integration.md):
 - [Vegetation Index Analysis with
-  geospatialsuite](https://cwru-sdle.github.io/geospatialsuite/articles/vegetation-indices.md):
+  geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/vegetation-indices.md):
 - [Water Quality
-  Assessment](https://cwru-sdle.github.io/geospatialsuite/articles/water-quality.md):
+  Assessment](https://exelegch.github.io/geospatialsuite-docs/articles/water-quality.md):
 - [Complete Workflows and Case
-  Studies](https://cwru-sdle.github.io/geospatialsuite/articles/workflows-case-studies.md):
+  Studies](https://exelegch.github.io/geospatialsuite-docs/articles/workflows-case-studies.md):
