### 2026-01-02 #9536 @Exelegch
> [@kdorheim](https://github.com/kdorheim)
>
> I have opened some preliminary issues on the repository. While the package offers some impressive claims, I'm presently unable to check anything as running any meaningful examples is not possible without external data.
>
> Furthermore, the documentation is hard to navigate through in its present form. I have recommended that a pkgdown website be built so that the functionality can be better reviewed.
>
> [@Exelegch](https://github.com/Exelegch) please work through the issues I have posted for now as the documentation in its current form does not show what it does.

@benyamindsmith Thanks for the review. I have added all the concerns and responded through the issues.
