+++ R/05-mosaicing.R
@@ -100,24 +100,34 @@ create_raster_mosaic <- function(input_data, method = "merge", region_boundary =
 #'
 #' @description
 #' Intelligently select raster files that overlap with a specified region.
+#' For ASTER files (GDEM and WBD), uses filename-based coordinate extraction
+#' for fast filtering without loading full rasters.
 #'
 #' @param input_folder Directory containing raster files
 #' @param region_boundary Region boundary or bounding box
+#' @param overlap Logical. If TRUE (default), checks for actual overlap with region.
+#'   If FALSE, only includes tiles whose centroid falls within region.
+#'   Only applies to ASTER files; non-ASTER files always use overlap detection.
 #' @param buffer_size Buffer around region (in degrees)
 #'
 #' @return Character vector of relevant file paths
 #'
 #' @examples
 #' \donttest{
-#' # Select ASTER files for Michigan
+#' # Select ASTER files for Michigan using overlap detection
 #' michigan_files <- select_rasters_for_region("/aster/files", "Michigan")
+#'
+#' # Use centroid-based selection (faster, fewer tiles)
+#' michigan_files_centroid <- select_rasters_for_region("/aster/files", "Michigan", 
+#'                                                       overlap = FALSE)
 #'
 #' # Select with custom buffer
 #' nevada_files <- select_rasters_for_region("/data", "Nevada", buffer_size = 0.2)
 #' }
 #'
 #' @export
-select_rasters_for_region <- function(input_folder, region_boundary, buffer_size = 0.1) {
+select_rasters_for_region <- function(input_folder, region_boundary, 
+                                      overlap = TRUE, buffer_size = 0.1) {
 
   message("Selecting rasters for specified region...")
 
@@ -148,16 +158,34 @@ select_rasters_for_region <- function(input_folder, region_boundary, buffer_size
   for (file in all_files) {
     tryCatch({
       # Extract coordinates from ASTER filenames if applicable
-      if (grepl("ASTGTMV", basename(file))) {
-        coords <- extract_aster_coordinates(basename(file))
-        if (!is.null(coords)) {
-          if (coords$lat >= bbox_coords[2] && coords$lat <= bbox_coords[4] &&
-              coords$lon >= bbox_coords[1] && coords$lon <= bbox_coords[3]) {
-            selected_files <- c(selected_files, file)
+      # Supports both GDEM (ASTGTMV) and WBD (ASTWBDV) formats
+      if (grepl("AST(GTM|WBD)V", basename(file))) {
+        
+        if (overlap) {
+          # Use bounding box overlap detection
+          aster_bbox <- extract_aster_coordinates(basename(file), boundingbox = TRUE)
+          if (!is.null(aster_bbox)) {
+            lon_overlap <- (aster_bbox$lon_max >= bbox_coords[1]) & 
+                          (bbox_coords[3] >= aster_bbox$lon_min)
+            lat_overlap <- (aster_bbox$lat_max >= bbox_coords[2]) & 
+                          (bbox_coords[4] >= aster_bbox$lat_min)
+            
+            if (unname(lon_overlap & lat_overlap)) {
+              selected_files <- c(selected_files, file)
+            }
+          }
+        } else {
+          # Use centroid-based selection (faster, fewer tiles)
+          coords <- extract_aster_coordinates(basename(file))
+          if (!is.null(coords)) {
[…]
