+++ articles/vegetation-indices.R
@@ -5,708 +5,335 @@ knitr::opts_chunk$set(
   fig.width = 8,
   fig.height = 6,
   warning = FALSE,
-  message = FALSE
+  message = FALSE,
+  eval = TRUE
 )
 
-## ----eval=FALSE---------------------------------------------------------------
-# library(geospatialsuite)
-# 
-# # See all available vegetation indices
-# all_indices <- list_vegetation_indices()
-# print(all_indices)
-# 
-# # View detailed information with formulas
-# detailed_indices <- list_vegetation_indices(detailed = TRUE)
-# head(detailed_indices)
-# 
-# # Filter by category
-# basic_indices <- list_vegetation_indices(category = "basic")
-# stress_indices <- list_vegetation_indices(category = "stress")
+## ----load-packages------------------------------------------------------------
+library(geospatialsuite)
+library(terra)
 
-## ----eval=FALSE---------------------------------------------------------------
-# # Simple NDVI from individual bands
-# ndvi <- calculate_vegetation_index(
-#   red = "landsat_red.tif",
-#   nir = "landsat_nir.tif",
-#   index_type = "NDVI",
+## ----quick-start--------------------------------------------------------------
+# Load sample spectral bands
+red <- load_sample_data("sample_red.rds")
+nir <- load_sample_data("sample_nir.rds")
+blue <- load_sample_data("sample_blue.rds")
+
+# Calculate NDVI using geospatialsuite
+ndvi <- calculate_vegetation_index(
+  red = red,
+  nir = nir,
+  index_type = "NDVI"
+)
+
+# Visualize
+plot(ndvi, main = "Normalized Difference Vegetation Index (NDVI)",
+     col = terrain.colors(100))
+
+## ----ndvi-example-------------------------------------------------------------
+# Calculate NDVI with geospatialsuite
+ndvi <- calculate_vegetation_index(
+  red = red,
+  nir = nir,
+  index_type = "NDVI"
+)
+
+# Summary statistics
+summary(values(ndvi))
+
+# Classify vegetation density
+vegetation_classes <- classify(ndvi,
+  rcl = matrix(c(-Inf, 0.2, 1,
+                 0.2, 0.6, 2,
+                 0.6, Inf, 3), 
+               ncol = 3, byrow = TRUE)
+)
+
+plot(vegetation_classes, 
+     main = "Vegetation Density Classes",
+     col = c("brown", "yellow", "darkgreen"),
+     legend = FALSE)
+legend("topright", 
+       legend = c("Sparse", "Moderate", "Dense"),
+       fill = c("brown", "yellow", "darkgreen"))
+
+## ----evi-example--------------------------------------------------------------
+# Calculate EVI using geospatialsuite
+evi <- calculate_vegetation_index(
+  red = red,
+  nir = nir,
+  blue = blue,
+  index_type = "EVI"
+)
+
+# Compare NDVI and EVI
+par(mfrow = c(1, 2))
+plot(ndvi, main = "NDVI", col = terrain.colors(100))
+plot(evi, main = "EVI", col = terrain.colors(100))
+par(mfrow = c(1, 1))
+
+## ----savi-example-------------------------------------------------------------
+# Calculate SAVI with geospatialsuite
+savi <- calculate_vegetation_index(
+  red = red,
+  nir = nir,
+  index_type = "SAVI"
+)
+
+plot(savi, main = "Soil Adjusted Vegetation Index (SAVI)",
+     col = terrain.colors(100))
+
+## ----multiple-indices---------------------------------------------------------
[…]
