+++ docs/reference/get_region_boundary.md
@@ -38,7 +38,7 @@ sf object with boundary geometry
 # \donttest{
 # US State with error handling
 ohio_boundary <- get_region_boundary("Ohio")
-#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   6%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  17%  |                                                                              |=================                                                     |  25%  |                                                                              |===================                                                   |  28%  |                                                                              |========================                                              |  34%  |                                                                              |=============================                                         |  41%  |                                                                              |=================================                                     |  48%  |                                                                              |=====================================                                 |  52%  |                                                                              |=========================================                             |  59%  |                                                                              |===========================================                           |  61%  |                                                                              |================================================                      |  69%  |                                                                              |=====================================================                 |  75%  |                                                                              |=============================================================         |  86%  |                                                                              |================================================================      |  91%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
+#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   6%  |                                                                              |=========                                                             |  12%  |                                                                              |============                                                          |  17%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  28%  |                                                                              |========================                                              |  34%  |                                                                              |=============================                                         |  41%  |                                                                              |=================================                                     |  48%  |                                                                              |======================================                                |  55%  |                                                                              |===========================================                           |  61%  |                                                                              |===============================================                       |  68%  |                                                                              |====================================================                  |  74%  |                                                                              |=========================================================             |  81%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================| 100%
 
 # Custom bounding box with validation
 custom_area <- get_region_boundary(c(-84.5, 39.0, -82.0, 41.0))
