+++ docs/articles/workflows-case-studies.md
@@ -81,10 +81,10 @@ test_geospatialsuite_package_simple(verbose = TRUE)
     #> 
     #> 
     #> $test_output_dir
-    #> [1] "/tmp/RtmpXPnLDK"
+    #> [1] "/tmp/RtmpMj0JDG"
     #> 
     #> $timestamp
-    #> [1] "2026-01-06 10:39:48 EST"
+    #> [1] "2026-01-06 11:03:37 EST"
     #> 
     #> $test_approach
     #> [1] "simplified_robust"
