+++ docs/paper.md
@@ -1,0 +1,93 @@
+---
+title: 'geospatialsuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R'
+tags:
+- R
+- geospatial analysis
+- remote sensing
+- vegetation indices
+- NDVI
+- spatial interpolation
+- terra
+- sf
+- agricultural research
+- environmental monitoring
+date: "28 December 2025"
+output:
+  html_document:
+    df_print: paged
+  pdf_document: default
+bibliography: paper.bib
+authors:
+- name: Olatunde D. Akanbi
+  orcid: "0000-0001-7719-2619"
+  equal-contrib: true
+  affiliation: '1'
+  corresponding: true
+- name: Vibha Mandayam
+  orcid: "0009-0008-8628-9904"
+  equal-contrib: true
+  affiliation: '1'
+- name: Yinghui Wu
+  orcid: "0000-0003-3991-5155"
+  affiliation: '1'
+- name: Jeffrey Yarus
+  orcid: "0000-0002-9331-9568"
+  affiliation: '1'
+- name: Erika I. Barcelos
+  orcid: "0000-0002-9273-8488"
+  affiliation: '1'
+- name: Roger H. French
+  orcid: "0000-0002-6162-0532"
+  affiliation: '1'
+affiliations:
+- name: Case Western Reserve University, Cleveland, OH, USA
+  index: 1
+  ror: 051fd9666
+---
+
+
+# Summary
+
+A comprehensive R package for geospatiotemporal analysis, geospatialsuite provides researchers with a unified toolkit for multimodal spatial data integration. The package features over 60 vegetation indices with automatic satellite band detection, universal spatial join operations that work with any raster-vector combination, robust raster visualization with built-in error handling, and rapid mapping capabilities through the `quick_map()` function. Designed for reliability and universal applicability, geospatialsuite addresses critical needs in agricultural research, environmental monitoring, and remote sensing applications while maintaining computational efficiency for large datasets.
+
+The package seamlessly integrates with the modern R spatial ecosystem through `terra` [@hijmans2022terra] and `sf` [@pebesma2018simple], providing enhanced NDVI calculations with quality filtering, comprehensive water quality analysis using multiple indices (NDWI [@mcfeeters1996normalized], MNDWI [@xu2006modification], NDMI), crop data layer (CDL) analysis [@usda2024cdl], spatial interpolation techniques [@cressie1993statistics], and terrain analysis. Unlike existing solutions that are often limited by complex dependencies or geographic scope, geospatialsuite emphasizes universal functionality—working seamlessly across different regions, satellite platforms (Landsat, Sentinel-2, MODIS), and data types with comprehensive error handling throughout. The package has enabled research in agricultural systems monitoring [@akanbi2024multiscale] and multimodal data integration [@akanbi2024integrating].
+
+# Statement of Need
+
+Geospatial analysis in environmental and agricultural research requires reliable, standardized tools that can handle diverse datasets across different geographic regions and satellite platforms. While foundational packages like `terra` [@hijmans2022terra] and `sf` [@pebesma2018simple] provide essential spatial data handling capabilities, researchers often face significant challenges when attempting to integrate multiple data sources, calculate specialized indices, or create reliable visualizations from large raster datasets.
+
+Current solutions in the R ecosystem present several critical limitations: vegetation index packages typically cover only a subset of available indices and lack automatic band detection across different satellite platforms; spatial join operations often fail with edge cases or require extensive preprocessing; visualization tools frequently encounter memory issues with large rasters or produce inconsistent results; and multimodal data integration requires custom workflows that are difficult to reproduce, error-prone, and time-consuming to implement.
+
+geospatialsuite addresses these fundamental challenges through several key innovations:
+
[…]
