+++ docs/paper.bib
@@ -1,0 +1,166 @@
+@article{rouse1974monitoring,
+  title={Monitoring vegetation systems in the Great Plains with ERTS},
+  author={Rouse Jr, John W and Haas, RH and Schell, JA and Deering, DW},
+  journal={NASA special publication},
+  volume={351},
+  pages={309},
+  year={1974}
+}
+
+@article{huete2002overview,
+  title={Overview of the radiometric and biophysical performance of the MODIS vegetation indices},
+  author={Huete, Alfredo and Didan, Kamel and Miura, Tomoaki and Rodriguez, E Patricia and Gao, Xiang and Ferreira, Laerte G},
+  journal={Remote Sensing of Environment},
+  volume={83},
+  number={1-2},
+  pages={195--213},
+  year={2002},
+  publisher={Elsevier},
+  doi={10.1016/S0034-4257(02)00096-2}
+}
+
+@article{mcfeeters1996normalized,
+  title={The use of the Normalized Difference Water Index (NDWI) in the delineation of open water features},
+  author={McFeeters, Stuart K},
+  journal={International Journal of Remote Sensing},
+  volume={17},
+  number={7},
+  pages={1425--1432},
+  year={1996},
+  publisher={Taylor \& Francis},
+  doi={10.1080/01431169608948714}
+}
+
+@article{gao1996ndwi,
+  title={NDWI—A normalized difference water index for remote sensing of vegetation liquid water from space},
+  author={Gao, Bo-cai},
+  journal={Remote Sensing of Environment},
+  volume={58},
+  number={3},
+  pages={257--266},
+  year={1996},
+  publisher={Elsevier},
+  doi={10.1016/S0034-4257(96)00067-3}
+}
+
+@book{cressie1993statistics,
+  title={Statistics for Spatial Data},
+  author={Cressie, Noel AC},
+  year={1993},
+  publisher={John Wiley \& Sons},
+  address={New York},
+  isbn={9780471002556}
+}
+
+@article{hijmans2022terra,
+  title={terra: Spatial Data Analysis},
+  author={Hijmans, Robert J},
+  journal={R package version 1.6-17},
+  year={2022},
+  url={https://CRAN.R-project.org/package=terra}
+}
+
+@article{pebesma2018simple,
+  title={Simple features for {R}: standardized support for spatial vector data},
+  author={Pebesma, Edzer},
+  journal={The R Journal},
+  volume={10},
+  number={1},
+  pages={439--446},
+  year={2018},
+  doi={10.32614/RJ-2018-009}
+}
+
+@article{wickham2016ggplot2,
+  title={ggplot2: Elegant Graphics for Data Analysis},
+  author={Wickham, Hadley},
+  year={2016},
+  publisher={Springer-Verlag New York},
+  isbn={978-3-319-24277-4},
+  url={https://ggplot2.tidyverse.org}
+}
+
+@Manual{usda2024cdl,
+  title={Cropland Data Layer},
+  author={{USDA National Agricultural Statistics Service}},
+  year={2024},
+  url={https://www.nass.usda.gov/Research_and_Science/Cropland/},
+  note={Accessed: 2024-12-01}
+}
+
+@article{gordon2025geospatial,
+  title={Geospatial modeling of near subsurface temperatures of the contiguous United States for assessment of materials degradation},
+  author={Gordon, Jonathan E and Akanbi, Olatunde D and Bhuvanagiri, Deepa C and Omodolor, Hope E and Mandayam, Vibha and French, Roger H and Yarus, Jeffrey M and Barcelos, Erika I},
+  journal={Scientific Reports},
+  volume={15},
+  number={1},
+  pages={1053},
+  year={2025},
[…]
