### 2026-01-14 #9536 @kthyng
@Exelegch I have read back through your comments on your short public history and I realize that I may have missed an important point. It sounds like you continue to develop in the private repository and periodically push updates to the open version on github. Is this correct? If so, it does not meet our mission as the Journal for Open Source Software nor does it allow for community development if users are not able to interact with the "real" version of the repository.
