### 2026-01-14 #9536 @Exelegch
> [@Exelegch](https://github.com/Exelegch) I have read back through your comments on your short public history and I realize that I may have missed an important point. It sounds like you continue to develop in the private repository and periodically push updates to the open version on github. Is this correct? If so, it does not meet our mission as the Journal for Open Source Software nor does it allow for community development if users are not able to interact with the "real" version of the repository.

@kthyng Thanks for flagging this — let me clarify.

All active development, issue tracking, and releases happen on this public GitHub repository, which is the canonical source of the software.

The project originated as an NSF research codebase (starting in 2022) maintained privately before it was refactored into a package. When we released v0.1.0 and submitted to CRAN, development moved to GitHub and has continued here through v0.1.1.

The Bitbucket repository now serves only as a mirror/archival copy for the NSF project record and is updated from GitHub; it is not used for development and does not contain features beyond what is publicly available here.

I apologize if my earlier comments suggested otherwise.
