### 2026-01-14 #9536 @Exelegch
@kthyng Let me make some thing clearer...

Thank you for your patience — I see the source of the confusion now and would like to clarify the timeline explicitly.

What I described on Dec 2 regarding private, parallel development was accurate for the pre-package research phase of this NSF project, before the software was released publicly. That is no longer the case.

Since the first public package release (v0.1.0, CRAN/GitHub), all active development has occurred on GitHub, which is the canonical and most up-to-date repository. The Bitbucket repository now serves only as an internal mirror for project records and is behind GitHub (last mirrored on Dec 11, 2025).

Regarding the fork, I only acted based on this response:

_

> Hi Olatunde,
>
> You recently submitted to JOSS but without a github handle (https://github.com/openjournals/joss-reviews/issues/9027). Can you resubmit but include your github handle?
>
> Thanks,
> Kristen

_
 I understand now that it was unnecessary. I am happy to proceed with cwru-sdle/geospatialsuite as the single canonical repository and archive or deprecate my personal fork to avoid any ambiguity. Let me know what to do...
