### 2026-01-16 cwru-sdle/geospatialsuite#1 @Exelegch
JOSS review [code]: devtools::test warnings #4 (original issue at https://github.com/Exelegch/geospatialsuite/issues/4)
[benyamindsmith](https://github.com/benyamindsmith)
opened [3 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/4#issue-3763965813)
After cloning the repository, I have ran tests with devtools::test() and got the following result:

[ FAIL 0 | WARN 2 | SKIP 0 | PASS 68 ]

Warning (test-integration.R:55:7): workflow integration works
Index EVI produced only NA values - skipping
*Warning (test-workflows.R:18:3): NDVI workflow completes successfully
Index EVI produced only NA values - skipping
Please explain these warnings and offer a remedy to ensure that they run successfully.

Activity
Exelegch
Exelegch commented 2 weeks ago
[Exelegch](https://github.com/Exelegch)
[2 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/4#issuecomment-3704260840)
Owner
Thanks for this... This is because we don't supply the needed band...

Fixed by adding blue band to test fixtures
All tests now pass cleanly
benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/4#issuecomment-3715319556) · edited by [benyamindsmith](https://github.com/benyamindsmith)
Author
@Exelegch

I just ran devtools::test() on the package and I got this traceback:

Warning (test-integration.R:55:7): workflow integration works
Failed to calculate EVI: blue band is required but not provided
Backtrace:
     ▆
  1. ├─testthat::expect_no_error(...) at test-integration.R:53:3
  2. │ └─testthat:::expect_no_(...)
  3. │   └─testthat:::quasi_capture(enquo(object), NULL, capture)
  4. │     ├─testthat (local) .capture(...)
  5. │     │ └─rlang::try_fetch(...)
  6. │     │   ├─base::tryCatch(...)
  7. │     │   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
  8. │     │   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
  9. │     │   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 10. │     │   └─base::withCallingHandlers(...)
 11. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
 12. ├─base::suppressMessages(...) at test-integration.R:54:5
 13. │ └─base::withCallingHandlers(...)
 14. └─geospatialsuite::run_comprehensive_geospatial_workflow(config) at test-integration.R:55:7
 15.   ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:63:3
 16.   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 17.   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 18.   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 19.   └─geospatialsuite::run_enhanced_ndvi_crop_workflow(...) at geospatialsuite/R/12-workflows.R:68:14
 20.     ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:214:5
 21.     │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 22.     │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 23.     │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 24.     └─geospatialsuite::calculate_multiple_indices(...) at geospatialsuite/R/12-workflows.R:215:7
 25.       └─base::tryCatch(...) at geospatialsuite/R/02-vegetation-indices.R:1067:7
 26.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 27.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 28.             └─value[[3L]](cond)

Warning (test-integration.R:55:7): workflow integration works
Failed to apply region boundary: region_boundary must be either an sf object, file path, or region name. Returning uncropped results.
Backtrace:
     ▆
  1. ├─testthat::expect_no_error(...) at test-integration.R:53:3
  2. │ └─testthat:::expect_no_(...)
  3. │   └─testthat:::quasi_capture(enquo(object), NULL, capture)
  4. │     ├─testthat (local) .capture(...)
  5. │     │ └─rlang::try_fetch(...)
  6. │     │   ├─base::tryCatch(...)
  7. │     │   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
  8. │     │   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
  9. │     │   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 10. │     │   └─base::withCallingHandlers(...)
 11. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
 12. ├─base::suppressMessages(...) at test-integration.R:54:5
 13. │ └─base::withCallingHandlers(...)
 14. └─geospatialsuite::run_comprehensive_geospatial_workflow(config) at test-integration.R:55:7
 15.   ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:63:3
 16.   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 17.   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 18.   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 19.   └─geospatialsuite::run_enhanced_ndvi_crop_workflow(...) at geospatialsuite/R/12-workflows.R:68:14
 20.     ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:214:5
 21.     │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 22.     │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 23.     │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 24.     └─geospatialsuite::calculate_multiple_indices(...) at geospatialsuite/R/12-workflows.R:215:7
 25.       └─base::tryCatch(...) at geospatialsuite/R/02-vegetation-indices.R:1104:5
 26.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 27.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 28.             └─value[[3L]](cond)
─────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          5 | mapping
✔ |          5 | package-functions
✔ |          4 | spatial-operations
⠏ |          0 | validation
Rasters have different geometries. Attempting to align...
✔ |          6 | validation
✔ |         11 | vegetation-indices
✔ |          5 | water-indices
⠏ |          0 | workflows
Starting comprehensive geospatial workflow with simplified visualization...
Starting enhanced NDVI crop analysis workflow...
Step 1: Calculating enhanced NDVI...
Starting NDVI calculation for time series analysis...
Using provided SpatRaster objects directly
NDVI calculation completed successfully!
  - 1 layer(s) created
  - Value range: [0.201, 0.760]
Step 2: Calculating additional vegetation indices...
⠋ |   1      0 | workflows
Step 3: Applying region boundary...
Step 5: Creating visualizations...
Step 6: Saving raster results...
Step 7: Generating statistics...
Enhanced NDVI crop workflow completed successfully!
Comprehensive workflow completed!
Results saved to: C:\Users\ben29\AppData\Local\Temp\Rtmp8mgwok
✔ |   2      1 | workflows
─────────────────────────────────────────────────────────────────────────────────────────────────
Warning (test-workflows.R:18:3): NDVI workflow completes successfully
Failed to calculate EVI: blue band is required but not provided
Backtrace:
     ▆
  1. ├─testthat::expect_no_error(workflow_result <- run_comprehensive_geospatial_workflow(config)) at test-workflows.R:18:3
  2. │ └─testthat:::expect_no_(...)
  3. │   └─testthat:::quasi_capture(enquo(object), NULL, capture)
  4. │     ├─testthat (local) .capture(...)
  5. │     │ └─rlang::try_fetch(...)
  6. │     │   ├─base::tryCatch(...)
  7. │     │   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
  8. │     │   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
  9. │     │   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 10. │     │   └─base::withCallingHandlers(...)
 11. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
 12. └─geospatialsuite::run_comprehensive_geospatial_workflow(config)
 13.   ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:63:3
 14.   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 15.   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 16.   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 17.   └─geospatialsuite::run_enhanced_ndvi_crop_workflow(...) at geospatialsuite/R/12-workflows.R:68:14
 18.     ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:214:5
 19.     │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 20.     │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 21.     │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 22.     └─geospatialsuite::calculate_multiple_indices(...) at geospatialsuite/R/12-workflows.R:215:7
 23.       └─base::tryCatch(...) at geospatialsuite/R/02-vegetation-indices.R:1067:7
 24.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 25.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 26.             └─value[[3L]](cond)

Warning (test-workflows.R:18:3): NDVI workflow completes successfully
Failed to apply region boundary: region_boundary must be either an sf object, file path, or region name. Returning uncropped results.
Backtrace:
     ▆
  1. ├─testthat::expect_no_error(workflow_result <- run_comprehensive_geospatial_workflow(config)) at test-workflows.R:18:3
  2. │ └─testthat:::expect_no_(...)
  3. │   └─testthat:::quasi_capture(enquo(object), NULL, capture)
  4. │     ├─testthat (local) .capture(...)
  5. │     │ └─rlang::try_fetch(...)
  6. │     │   ├─base::tryCatch(...)
  7. │     │   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
  8. │     │   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
  9. │     │   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 10. │     │   └─base::withCallingHandlers(...)
 11. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
 12. └─geospatialsuite::run_comprehensive_geospatial_workflow(config)
 13.   ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:63:3
 14.   │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 15.   │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 16.   │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 17.   └─geospatialsuite::run_enhanced_ndvi_crop_workflow(...) at geospatialsuite/R/12-workflows.R:68:14
 18.     ├─base::tryCatch(...) at geospatialsuite/R/12-workflows.R:214:5
 19.     │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 20.     │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 21.     │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
 22.     └─geospatialsuite::calculate_multiple_indices(...) at geospatialsuite/R/12-workflows.R:215:7
 23.       └─base::tryCatch(...) at geospatialsuite/R/02-vegetation-indices.R:1104:5
 24.         └─base (local) tryCatchList(expr, classes, parentenv, handlers)
 25.           └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
 26.             └─value[[3L]](cond)
─────────────────────────────────────────────────────────────────────────────────────────────────

══ Results ══════════════════════════════════════════════════════════════════════════════════════
Duration: 4.9 s

[ FAIL 0 | WARN 4 | SKIP 0 | PASS 68 ]
[Exelegch](https://github.com/Exelegch)
Exelegch commented [last week](https://github.com/Exelegch/geospatialsuite/issues/4#issuecomment-3715448914)
Exelegch
last week
Owner
@benyamindsmith Oh! I was working on the version you checked based on you suggestion. It seems we were working at the same time. It should be fine now

benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/4#issuecomment-3720972040)
Author
@Exelegch just ran the tests and everything checks out! You can close this issue.

[kdorheim](https://github.com/kdorheim)
mentioned this [4 days ago](https://github.com/Exelegch/geospatialsuite/issues/4#event-4345730520)
[[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R openjournals/joss-reviews#9536](https://github.com/openjournals/joss-reviews/issues/9536)

[Exelegch](https://github.com/Exelegch)
closed this as [completed](https://github.com/Exelegch/geospatialsuite/issues?q=is%3Aissue%20state%3Aclosed%20archived%3Afalse%20reason%3Acompleted)[3 days ago](https://github.com/Exelegch/geospatialsuite/issues/4#event-21990922832)
