### 2026-01-16 cwru-sdle/geospatialsuite#2 @Exelegch
JOSS review [documentation]: Unable to run any meaningful examples in documentation #3 (Original issue at https://github.com/Exelegch/geospatialsuite/issues/3)
[benyamindsmith](https://github.com/benyamindsmith)
opened [3 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/3#issue-3763963356)
I have tried to run some of the examples in the vingettes and in the README and I haven't had much success with actually doing anything the paper claims the package does.

Please include and/or link datasets in the package that will run the code. If adding too many datasets will increase bloat.

Please make a function that will allow the user to download different datasets for the examples.

Activity
Exelegch
Exelegch commented 2 weeks ago
[Exelegch](https://github.com/Exelegch)
[2 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/3#issuecomment-3704259622)
Owner
Added list_sample_datasets() and load_sample_data() functions for instant testing
All vignettes and README examples now use built-in sample data
Users can run examples immediately after installation without additional downloads
Implemented via load_sample_data() - provides sample rasters (red, NIR, green, blue, SWIR) and vector data without package bloat
benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/3#issuecomment-3715313650)
Author
@Exelegch

Thanks for doing this. If the examples run quickly, can you please ensure that the examples the function documentation run as well?

I notice that the example functions are wrapped with \dontrun{} when I think it would be prudent to allow for the user to run them.

See here for why I think this is prefered: https://stackoverflow.com/questions/68935126/when-and-when-not-to-use-donttest-and-dontrun-in-r-package-development

Exelegch
Exelegch commented last week
[Exelegch](https://github.com/Exelegch)
[last week](https://github.com/Exelegch/geospatialsuite/issues/3#issuecomment-3715545290)
Owner
Thanks @benyamindsmith We appreciate this feedback and have carefully reviewed our example documentation.
We want to clarify our current approach:
Our package already uses both \donttest{} and \dontrun{} appropriately based on R best practices:

Functions with \donttest{} (can run, but skipped in checks):
These examples use our built-in sample data and will execute if users run them:
calculate_vegetation_index() - with sample raster bands
load_sample_data() - demonstrates data loading
list_sample_datasets() - lists available data
universal_spatial_join() - with sample points
(~10-12 functions already use \donttest{} with sample data)

Functions that must keep \dontrun{} (cannot execute):
These require external resources that cannot be bundled with the package:
Example 1: Large external datasets
r#' https://github.com/examples
#' \dontrun{
#' # Requires ASTER DEM files (100+ GB datasets not included)
#' michigan_files <- select_rasters_for_region("/aster/files", "Michigan")
#'
#' # Requires user's own satellite imagery
#' ndvi <- calculate_vegetation_index(
#' spectral_data = "/path/to/sentinel/bands/",
#' index_type = "NDVI"
#' )
#' }
Example 2: External services/APIs
r#' https://github.com/examples
#' \dontrun{
#' # Requires tigris package and Census API access
#' boundary <- get_region_boundary("Ohio")
#'
#' # Requires user's directory structure
#' mosaic <- create_raster_mosaic("/user/data/directory")
#' }
Why \dontrun{} is appropriate here:
Per [R documentation](https://cran.r-project.org/doc/manuals/R-exts.html#Documenting-functions) and the [StackOverflow discussion you linked](https://stackoverflow.com/questions/68935126/when-and-when-not-to-use-donttest-and-dontrun-in-r-package-development):
"Use \dontrun{} for code that should not be executed automatically, for example if it requires external resources unavailable during package checking."

As a geospatial analysis package, most of our functions are designed to work with:

Large satellite imagery datasets (Landsat, Sentinel-2, ASTER)
User-provided GIS files and shapefiles
External geospatial APIs and services

These datasets typically range from hundreds of megabytes to hundreds of gigabytes and cannot be included in a CRAN package.

Our Approach:
We provide small built-in sample data (~50-80 KB total in .rds format) for demonstrating core functionality, and these examples use \donttest{}. However, realistic geospatial workflows require external data, so those examples appropriately use \dontrun{} to show the intended usage without attempting execution.
We're happy to make improvements if you have specific suggestions!

benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/3#issuecomment-3715611150)
Author
@Exelegch thank you for clarifying!

You can close this issue now.

[Exelegch](https://github.com/Exelegch)
closed this as [completed](https://github.com/Exelegch/geospatialsuite/issues?q=is%3Aissue%20state%3Aclosed%20archived%3Afalse%20reason%3Acompleted)[last week](https://github.com/Exelegch/geospatialsuite/issues/3#event-21881406864)

[kdorheim](https://github.com/kdorheim)
mentioned this [4 days ago](https://github.com/Exelegch/geospatialsuite/issues/3#event-4345730519)
[[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R openjournals/joss-reviews#9536](https://github.com/openjournals/joss-reviews/issues/9536)
