### 2026-01-16 cwru-sdle/geospatialsuite#3 @Exelegch
JOSS review [documentation]: Make a pkgdown website #2 ( original issue at https://github.com/Exelegch/geospatialsuite/issues/2)
[benyamindsmith](https://github.com/benyamindsmith)
opened [3 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/2#issue-3763960121) · edited by [benyamindsmith](https://github.com/benyamindsmith)
While you do have documentation available on [CRAN](https://cran.r-project.org/web/packages/geospatialsuite/vignettes/) in its present form its difficult to navigate through quickly.

Given the scope of this package, it would pre prudent to create a [pkgdown website](https://pkgdown.r-lib.org/) for this package and have it associated with the repository.

Activity
Exelegch
Exelegch commented 2 weeks ago
[Exelegch](https://github.com/Exelegch)
[2 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/2#issuecomment-3704258250)
Owner
Thanks for this suggestion. Now live at https://exelegch.github.io/geospatialsuite-docs/ with full function reference, vignettes, and search functionality

benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/2#issuecomment-3714920133) · edited by [benyamindsmith](https://github.com/benyamindsmith)
Author
@Exelegch thanks for this. This looks great!

Please ensure the [Github actions workflow](https://github.com/Exelegch/geospatialsuite/deployments/github-pages) rebuilds the site when changes are pushed to the repository.

After that is taken care of, you can close this issue.

[benyamindsmith](https://github.com/benyamindsmith)
mentioned this [last week](https://github.com/Exelegch/geospatialsuite/issues/2#event-4298601264)
[JOSS review [documentation]: Package title consistency #1](https://github.com/Exelegch/geospatialsuite/issues/1)

[Exelegch](https://github.com/Exelegch)
closed this as [completed](https://github.com/Exelegch/geospatialsuite/issues?q=is%3Aissue%20state%3Aclosed%20archived%3Afalse%20reason%3Acompleted)[last week](https://github.com/Exelegch/geospatialsuite/issues/2#event-21881083627)

[kdorheim](https://github.com/kdorheim)
mentioned this [4 days ago](https://github.com/Exelegch/geospatialsuite/issues/2#event-4345730518)
[[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R openjournals/joss-reviews#9536](https://github.com/openjournals/joss-reviews/issues/9536)
