### 2026-01-16 cwru-sdle/geospatialsuite#4 @Exelegch
JOSS review [documentation]: Package title consistency #1 (Original issue at https://github.com/Exelegch/geospatialsuite/issues/1)
[benyamindsmith](https://github.com/benyamindsmith)
opened [3 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/1#issue-3763954533)
The package name is geospatialsuite (all lowercase), however when mentioned other documentation [PascalCase](https://khalilstemmler.com/img/blog/camel-snake-pascal-case/pascal-case.png) appears to be used.

I would recommend choosing one form and using it across all documentation

Activity
Exelegch
Exelegch commented 2 weeks ago
[Exelegch](https://github.com/Exelegch)
[2 weeks ago](https://github.com/Exelegch/geospatialsuite/issues/1#issuecomment-3704257690)
Owner
Thanks for this... We have standardized to lowercase "geospatialsuite" across all documentation, vignettes, README, and paper.md

benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/1#issuecomment-3714956756) · edited by [benyamindsmith](https://github.com/benyamindsmith)
Author
@Exelegch I just did a [case sensitive search](https://github.com/orgs/community/discussions/9759) of GeoSpatialSuite and still found alot of outstanding cases.

Please see my screenshot below:

Image
Also in the first word of the paper is capitalized.

Image
Maybe consider rewriting this sentence so you don't start the sentence with a lowercase letter.

Please ensure naming consistency across the package and paper.

(Also related to https://github.com/Exelegch/geospatialsuite/issues/2 but I assume that the pkgdown website will also be rebuilt via Github actions with these edits)

Exelegch
Exelegch commented last week
[Exelegch](https://github.com/Exelegch)
[last week](https://github.com/Exelegch/geospatialsuite/issues/1#issuecomment-3715441663)
Owner
@benyamindsmith Thanks for this... I have actually changed them. These files are cached files that were generated with the previous version (you will see they are four months old). But the viginettes and doc directory are the updated ones. I have moved them away and regenerated them. You should have updated files now with consistent names. Regarding the name in the paper, I struggled with it initially (First word must be capitalized), but i have adopted your suggestion and the package name is now embedded in the sentence.

And this updated in the Github action automatically

benyamindsmith
benyamindsmith commented last week
[benyamindsmith](https://github.com/benyamindsmith)
[last week](https://github.com/Exelegch/geospatialsuite/issues/1#issuecomment-3715447439)
Author
@Exelegch

Thanks for letting me know!

You can close this issue now.

[Exelegch](https://github.com/Exelegch)
closed this as [completed](https://github.com/Exelegch/geospatialsuite/issues?q=is%3Aissue%20state%3Aclosed%20archived%3Afalse%20reason%3Acompleted)[last week](https://github.com/Exelegch/geospatialsuite/issues/1#event-21881072898)

[kdorheim](https://github.com/kdorheim)
mentioned this [4 days ago](https://github.com/Exelegch/geospatialsuite/issues/1#event-4345730517)
[[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R openjournals/joss-reviews#9536](https://github.com/openjournals/joss-reviews/issues/9536)

[Exelegch](https://github.com/Exelegch)
mentioned this [1 minute ago](https://github.com/Exelegch/geospatialsuite/issues/1#event-4383431291)
[JOSS review [documentation]: Make a pkgdown website #2 ( original issue at https://github.com/Exelegch/geospatialsuite/issues/2) cwru-sdle/geospatialsuite#3](https://github.com/cwru-sdle/geospatialsuite/issues/3)
