### 2026-01-16 cwru-sdle/geospatialsuite#5 @Exelegch
JOSS review [functionality]: Performance Claims by quickmap() #5 (Original issue at https://github.com/Exelegch/geospatialsuite/issues/5)
[benyamindsmith](https://github.com/benyamindsmith)
opened [yesterday](https://github.com/Exelegch/geospatialsuite/issues/5#issue-3814550709)
In the JOSS paper it states:

Reliable visualization for large datasets: Built on terra's efficient raster handling, the package provides publication-quality mapping through functions like quick_map() that handle large rasters without memory issues, automatic legend generation, and consistent color schemes. The visualization system includes robust error handling that prevents common plotting failures.

This here is a performance claim which I am unable to find any verification for.

I have a the following questions/requests:

Please provide a benchmark to other mapping functions for similar datasets. A small table is efficient which compares memory usage and plotting time of quick_map() vs comparable functions in other R packages.
Please note comparable/related packages and their performance relative to geospatialsuite's quick_map() function
Please provide a reproducible code as well.
Activity

[benyamindsmith](https://github.com/benyamindsmith)
mentioned this [yesterday](https://github.com/Exelegch/geospatialsuite/issues/5#event-4372277874)
[[REVIEW]: GeoSpatialSuite: Comprehensive Geospatiotemporal Analysis and Multimodal Integration Toolkit for R openjournals/joss-reviews#9536](https://github.com/openjournals/joss-reviews/issues/9536)
