### 2026-01-16 ec2decd
Add performance benchmark addressing reviewer feedback
