+++ articles/performance_benchmark.R
@@ -1,0 +1,300 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  eval = FALSE  # Benchmarks require large external datasets not included in package
+)
+
+## ----benchmark-setup----------------------------------------------------------
+# library(geospatialsuite)
+# library(terra)
+# library(ggplot2)
+# library(bench)
+# 
+# # Simulate large raster dataset characteristics
+# # In practice, use real satellite imagery (Landsat, Sentinel-2, etc.)
+# create_test_raster <- function(nrow = 5000, ncol = 5000) {
+#   r <- rast(nrows = nrow, ncols = ncol,
+#             xmin = -180, xmax = 180,
+#             ymin = -90, ymax = 90)
+#   values(r) <- runif(ncell(r), 0, 1)
+#   names(r) <- "test_value"
+#   return(r)
+# }
+# 
+# # Create test datasets of varying sizes
+# cat("Creating test datasets (this may take a moment)...\n")
+# small_raster <- create_test_raster(1000, 1000)    # ~7.6 MB
+# medium_raster <- create_test_raster(3000, 3000)   # ~68.7 MB
+# large_raster <- create_test_raster(5000, 5000)    # ~190.7 MB
+# 
+# 
+
+## ----geospatialsuite-method---------------------------------------------------
+# plot_geospatialsuite <- function(raster_data) {
+#   quick_map(raster_data, title = "geospatialsuite Plot")
+# }
+
+## ----terra-method-------------------------------------------------------------
+# plot_terra_base <- function(raster_data) {
+#   plot(raster_data, main = "terra Base Plot")
+# }
+
+## ----ggplot-method------------------------------------------------------------
+# plot_ggplot2 <- function(raster_data) {
+#   # Convert raster to data frame (this is what causes memory scaling)
+#   df <- as.data.frame(raster_data, xy = TRUE)
+#   ggplot(df, aes(x = x, y = y, fill = test_value)) +
+#     geom_raster() +
+#     scale_fill_viridis_c() +
+#     theme_minimal() +
+#     labs(title = "ggplot2 Plot")
+# }
+
+## ----memory-benchmark---------------------------------------------------------
+# 
+# 
+# test_datasets <- list(
+#   Small = small_raster,
+#   Medium = medium_raster,
+#   Large = large_raster
+# )
+# 
+# cat("\n=== MEMORY FOOTPRINT ANALYSIS ===\n\n")
+# 
+# memory_summary <- data.frame(
+#   Dataset = character(),
+#   Dimensions = character(),
+#   Pixels = character(),
+#   ggplot2_MB = numeric(),
+#   terra_geospatial_MB = character(),
+#   stringsAsFactors = FALSE
+# )
+# 
+# for (dataset_name in names(test_datasets)) {
+#   raster_data <- test_datasets[[dataset_name]]
+#   dims <- sprintf("%d x %d", nrow(raster_data), ncol(raster_data))
+#   n_pixels <- format(ncell(raster_data), big.mark = ",")
+# 
+#   # The critical measurement: ggplot2's data frame size
+#   df <- as.data.frame(raster_data, xy = TRUE)
+#   df_size_mb <- as.numeric(object.size(df)) / 1024^2
+#   rm(df); gc(verbose = FALSE)
+# 
+# 
+#   if (df_size_mb > 75) {
+#     cat(sprintf("  Memory advantage:        %.1fx less memory\n\n", df_size_mb / 75))
[…]
