+++ docs/articles/index.md
@@ -8,6 +8,8 @@
   geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/analyze-crop-vegetation.md):
 - [Getting Started with
   geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/getting-started.md):
+- [Performance Benchmark: quick_map() vs Alternative
+  Approaches](https://exelegch.github.io/geospatialsuite-docs/articles/performance_benchmark.md):
 - [Universal Spatial Mapping with
   geospatialsuite](https://exelegch.github.io/geospatialsuite-docs/articles/spatial-analysis.md):
 - [Spatial Data
