+++ docs/articles/workflows-case-studies.md
@@ -74,17 +74,17 @@ test_geospatialsuite_package_simple(verbose = TRUE)
     #> [1] 100
     #> 
     #> $summary$duration_seconds
-    #> [1] 0.29
+    #> [1] 0.34
     #> 
     #> $summary$version
     #> [1] "0.1.0"
     #> 
     #> 
     #> $test_output_dir
-    #> [1] "/tmp/Rtmpvvm01C"
+    #> [1] "/tmp/RtmpovOQmj"
     #> 
     #> $timestamp
-    #> [1] "2026-01-06 11:42:12 EST"
+    #> [1] "2026-01-16 11:09:38 EST"
     #> 
     #> $test_approach
     #> [1] "simplified_robust"
