### 2026-01-16 c14ca55
Add performance benchmark addressing reviewer feedback2
