### 2026-01-19 #9536 @benyamindsmith
@kdorheim

I can confirm that @Exelegch addressed all my issues. `geospatialsuite` offers a "one stop shop" of tools for geospatial scientists especially those specializing in the area of environmental/climate science.

Its important to note that my review focused primarily on the functional aspects and the package structure. While I do understand the package's utility, I would highly recommend another subject matter expert to review the material.

I'm happy with my review on my side. I recommend that JOSS should accept `geospatialsuite` as a submission.
