### 2026-02-02 #9536 @jjrob
## Important Note

It appears that the reviewer guidelines and checklist changed after this submission was received and reviews started. On January 7, @arfon instructed that this review should occur under the previous guidelines. However, it appears the @editorialbot generated the new checklist and doesn't have a command to generate the old one. Unless anyone objects, I'm going to use the new one, except with regard to AI usage. The previous guidelines only said the following about AI:

> JOSS does not have a policy regarding the use of AI/LLMs yet, but authors are responsible for understanding and explaining submitted code and its provenance, and should respond in good faith to reviewer questions about LLM use as they would with any other topic.

I'm not going to examine the submission in detail for AI/LLM usage, but if I notice anything I might ask about it.

For my part, I'm not going to use AI / LLMs in any way during my review, and will go ahead and check the checkbox under **AI usage** below.

****
