### 2026-02-02 cwru-sdle/geospatialsuite#6 @jjrob
The list of dependencies in README.Rmd do not match each other, or the DESCRIPTION file
In README.Rmd, the **Installation** section at the top lists "Recommended Packages for Best Performance". This list of packages is shorter than that in the **System Requirements** section at the bottom, which lists "Recommended Packages for Enhanced Functionality". I suggest you synchronize these two sections. The two sections should list the same packages and use the same descriptions (i.e. use the same term, e.g. "Best Performance" or "Enhanced Functionality", in both places). Or you could remove one section, or consolidate the two of them, to avoid having to maintain two lists.

Also, please cross-check the dependencies listed in the README with the Imports and Suggests lists in the DESCRIPTION file to make sure they are sufficiently consistent. E.g. DESCRIPTION lists leaflet as an import (so it will be installed with geospatialsuite is installed) but README lists it as "recommended" (implying it is optional).

This issue was raised as part of openjournals/joss-reviews#9536.
