### 2026-02-02 cwru-sdle/geospatialsuite#7 @jjrob
In paper.bib, a few references are missing URLs/DOIs or have the wrong one
Please check these:

1. "Leveraging Multimodal Data for Geospatiotemporal Analytics". Please add a link to your thesis if your university makes it available online.
2. "Integrating Multimodal Geospatiotemporal Data for Societal, Economic, and Environmental (SEE) Analysis of Large Agricultural Systems" has the wrong DOI linked (https://doi.org/10.1109/BigData62323.2024.10825456). Please fix it.
3. "Monitoring Vegetation Systems in the Great Plains with ERTS". If possible, please add link to NASA's canonical location for this document, e.g. https://ntrs.nasa.gov/citations/19740022614

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
