### 2026-02-02 cwru-sdle/geospatialsuite#8 @jjrob
The rendered paper.pdf in the repo does not list the authors
paper.md lists the authors in the correct metadata format but the rendered PDF does not show them. Should it?

I'm not sure if this is problem with what you have done, or is just how JOSS publishing works. I could not determine if the JOSS authorship guidelines required you to include authors in a way that would be rendered (vs. just in the metadata), or if those guidelines explicitly said NOT to do that. In any case, it is confusing to have this pre-print pdf in the repo without authors. It seems possible that search engines will index it and that someone searching for the paper would come across the pre-print, which would be confusing for them if no authors were listed.

I suggest you consult with the JOSS editors about what is appropriate. Feel free to close this issue as if everything has actually been done correctly.

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
