### 2026-02-02 cwru-sdle/geospatialsuite#9 @jjrob
In paper.md, the Summary section may not be sufficiently approachable by a "diverse, non-specialist audience"
The JOSS paper guidelines say:

> A summary describing the high-level functionality and purpose of the software for a diverse, *non-specialist audience.* [emphasis in original]

As currently written, the Summary section is quite specialized and technical. I support this—as a specialist in an adjacent field, I appreciate the technical detail—but in light of JOSS's request that the Summary be oriented toward non-specialists, I suggest you add an introductory sentence designed for them. Your first sentence under Statement of Need would be great; perhaps you could move it from the Statement of Need to the Summary.

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
