### 2026-02-02 cwru-sdle/geospatialsuite#10 @jjrob
Remove "includes auto-geocoding!" from README
The README includes the code block:

```R
# Install development version with latest features (includes auto-geocoding!)
# install.packages("devtools")
devtools::install_github("cwru-sdle/geospatialsuite")
```

I humbly suggest you remove the `(includes auto-geocoding!)` from the comment. Although that may be an exciting feature, this statement introduces tension with the heading above, which says `From CRAN (Recommended)`. Now the user is under stress: should they take the recommendation to install from CRAN and forgo the new exciting feature? Or do they want the dev version to get that feature?

In fact, it appears that auto-geocoding is actually part of the released version, or it will be soon. So you might as well remove `(includes auto-geocoding!)`. I humbly suggest that it will be tedious for you to maintain your README comments to note every new feature coming up, and then remove them as builds are released. A more traditional approach would be to create a ROADMAP file, or a RELEASE_NOTES or CHANGES file that you update with new features and fixes as they are made.

This is a very minor issue—just a suggestion—so if you don't agree feel free to close it!

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
