### 2026-02-02 cwru-sdle/geospatialsuite#11 @jjrob
geospatialsuite-package.Rd and DESCRIPTION do not list the same authors
man/geospatialsuite-package.Rd, which shows up in the R package documentation, only lists three authors, while DESCRIPTION, which shows up in CRAN and other places, lists six, which are the same as are listed for the paper. Should geospatialsuite-package.Rd be updated to contain all six authors?

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
