### 2026-02-02 cwru-sdle/geospatialsuite#12 @jjrob
The README references CODE_OF_CONDUCT.md, but it doesn't exist
The README says:

> Code of Conduct
>
> This project adheres to a [Code of Conduct](https://github.com/cwru-sdle/geospatialsuite/blob/main/CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code.

But that CODE_OF_CONDUCT.md file does not exist. You should either create it or remove the mention of a Code of Conduct from the README. (Note: I have no opinion on whether you should have a Code of Conduct, what should be in it, whether JOSS requires one as a condition of publishing, and if they do, what they might insist upon. It would probably be best to check with the JOSS editor.)

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
