### 2026-02-02 cwru-sdle/geospatialsuite#13 @jjrob
The README "Quick Start" steps do not work for the geospatialsuite package 0.1.1 installed from CRAN
In the README, under Quick Start --> Basic Usage Examples, the code starts with:

```R
# Load built-in sample data
red <- load_sample_data("sample_red.rds")
nir <- load_sample_data("sample_nir.rds")
blue <- load_sample_data("sample_blue.rds")
```

This fails with geospatialsuite 0.1.1 installed from CRAN: `load_sample_data()` does not exist in that package.

I suspect what happened is that you added `load_sample_data()` after releasing 0.1.1 to CRAN, perhaps in response to prior feedback during this JOSS review, but you did not update the package version in the github repo. This makes it seem like the version on CRAN, 0.1.1 is the latest and the same as what's in github, when in fact the dev version in github has additional functionality but the same version number.

I suggest you update the version number in the github repo to 0.2.0, under the assumption that you are using [semantic versioning](https://semver.org/) and that you've added new functionality rather than just fixed bugs. (If you just fixed bugs, then 0.1.2 would be appropriate, but I don't think you did that.)

Then, in the future, I suggest you increment the version number in the git repo as part of the first commit you make after issuing a release. That way, as soon as the git repo advances from CRAN, it has a newer version number.

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
