### 2026-02-02 cwru-sdle/geospatialsuite#14 @jjrob
In the README, not all of the Quick Start --> Basic Usage Examples and Auto-Geocoding Without Coordinates examples actually work
### Basic Usage Examples

Some of the Basic Usage Examples reference non-existing data or parameters. Can you please check all of the examples and ensure they work "right out of the box" after the user installs the package?

This works:

```R
> red <- load_sample_data("sample_red.rds")
> nir <- load_sample_data("sample_nir.rds")
> blue <- load_sample_data("sample_blue.rds")
```

This fails:

```R
> quick_map("mydata.shp")
Error: Could not process spatial data
```

This fails:

```R
> result <- universal_spatial_join(
+     source_data = "study_sites.csv",
+     target_data = "elevation.tif",
+     method = "extract"
+ )
Error: Failed to process source_data: File or directory does not exist: study_sites.csv
```

This works:

```R
> ndvi <- calculate_vegetation_index(
+     red = red,
+     nir = nir,
+     index_type = "NDVI",
+     verbose = TRUE
+ )
Starting NDVI calculation with comprehensive input handling...
Loading and validating input bands with CRS checking...
Checking and fixing spatial compatibility...
Computing NDVI index...
NDVI calculation completed:
  - Valid pixels: 100/100 (100.0%)
  - Value range: [0.478454, 0.875212]
  - Mean: 0.706874, Std Dev: 0.084458
  - Percentiles (25th, 50th, 75th): [0.6473, 0.7172, 0.7725]
```

This fails:

```R
> enhanced_ndvi <- calculate_ndvi_enhanced(
+     red = red,
+     nir = nir,
+     quality_filter = TRUE,
+     mask_invalid = TRUE
+ )
Error in calculate_ndvi_enhanced(red = red, nir = nir, quality_filter = TRUE,  :
  unused argument (mask_invalid = TRUE)
```

This works:

```R
> indices <- calculate_multiple_indices(
+     red = red,
+     nir = nir,
+     blue = blue,
+     indices = c("NDVI", "EVI", "SAVI"),
+     output_stack = TRUE
+ )
```

### Auto-Geocoding Without Coordinates

Similarly, not all of these work. Can you please ensure they do? If the user needs to install a package for them to work, please say so in the README. (You can even show the installation of it in code.)

This fails:

```R
> state_data <- data.frame(
+     state = c("Ohio", "PA", "Michigan"),
+     population = c(11.8, 13.0, 10.1)
+ )
> spatial_data <- auto_geocode_data(state_data)
No coordinate columns found. Attempting to detect geographic entities...
Error: Could not find coordinate columns or geographic entities to geocode.
Available columns: state, population

Supported geographic entities: state, county, FIPS, HUC, ZIP code, city
```

This fails:

```R
> county_data <- data.frame(
+     fips = c("39049", "39035", "39113"),
+     unemployment = c(4.2, 5.1, 4.8)
+ )
> county_sf <- auto_geocode_data(county_data)
No coordinate columns found. Attempting to detect geographic entities...
Error: Could not find coordinate columns or geographic entities to geocode.
Available columns: fips, unemployment

Supported geographic entities: state, county, FIPS, HUC, ZIP code, city
```

This fails but directs the user to install a required package, which is fine:

```R
> watershed_data <- data.frame(
+     HUC_8 = c("04100009", "04100012"),  # or HUC-8, huc8, Huc 8, etc.
+     water_quality = c(72, 65)
+ )
> huc_sf <- auto_geocode_data(watershed_data)
No coordinate columns found. Attempting to detect geographic entities...
Detected huc data in column: HUC_8
Geocoding HUC codes...
Error: Package 'nhdplusTools' is required for HUC geocoding.
Install it with: install.packages('nhdplusTools')
```
But then, when I try it again after installing the package, this happens:

```R
> install.packages('nhdplusTools')
Installing package into ‘C:/jjr8/JOSS_test/RLibDir’
(as ‘lib’ is unspecified)
also installing the dependencies ‘timechange’, ‘data.table’, ‘RANN’, ‘lubridate’, ‘R.oo’, ‘R.methodsS3’, ‘maplegend’, ‘fstcore’, ‘assertthat’, ‘hydroloom’, ‘dataRetrieval’, ‘R.utils’, ‘maptiles’, ‘mapsf’, ‘fst’, ‘arrow’, ‘pbapply’
...
> huc_sf <- auto_geocode_data(watershed_data)
No coordinate columns found. Attempting to detect geographic entities...
Detected huc data in column: HUC_8
Geocoding HUC codes...
Detected HUC level: huc08
defaulting to 2020 version of WBD
defaulting to 2020 version of WBD
Error in UseMethod("st_centroid") :
  no applicable method for 'st_centroid' applied to an object of class "data.frame"
```

Loading `sf` does not fix it:

```R
> library(sf)
Linking to GEOS 3.13.1, GDAL 3.11.4, PROJ 9.7.0; sf_use_s2() is TRUE
> huc_sf <- auto_geocode_data(watershed_data)
No coordinate columns found. Attempting to detect geographic entities...
Detected huc data in column: HUC_8
Geocoding HUC codes...
Detected HUC level: huc08
defaulting to 2020 version of WBD
defaulting to 2020 version of WBD
Error in UseMethod("st_centroid") :
  no applicable method for 'st_centroid' applied to an object of class "data.frame"
```

A similar experience here, except that it works after installing the package:

```R
> zip_data <- data.frame(
+     zip = c("43215", "44113", "45202"),
+     median_income = c(58000, 45000, 72000)
+ )
> zip_sf <- auto_geocode_data(zip_data)
No coordinate columns found. Attempting to detect geographic entities...
Detected zipcode data in column: zip
Geocoding ZIP codes...
Error: Package 'zipcodeR' is required for ZIP code geocoding.
Install it with: install.packages('zipcodeR')

> install.packages('zipcodeR')
Installing package into ‘C:/jjr8/JOSS_test/RLibDir’
(as ‘lib’ is unspecified)
also installing the dependencies ‘selectr’, ‘rvest’, ‘blob’, ‘plogr’, ‘tidycensus’, ‘RSQLite’
...

The downloaded binary packages are in
	C:\Users\jjr8\AppData\Local\Temp\2\Rtmp6h0OPL\downloaded_packages
> zip_sf <- auto_geocode_data(zip_data)
No coordinate columns found. Attempting to detect geographic entities...
Detected zipcode data in column: zip
Geocoding ZIP codes...
Matched 3/3 ZIP codes
```

This fails:

```R
> preview_geocoding(my_data)
Error: object 'my_data' not found
```

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
