### 2026-02-02 cwru-sdle/geospatialsuite#15 @jjrob
The README references `vignette("performance-benchmark")`, which fails in R
With the current version installed from github:

```R
> vignette("performance-benchmark")
Warning message:
vignette ‘performance-benchmark’ not found
```

Note that I did find the vignette manually in the documentation and did run all of the performance tests successfully. So good job with that! You just need to either fix this so `vignette("performance-benchmark")` works or remove the reference to it from the README.

This issue was raised as part of https://github.com/openjournals/joss-reviews/issues/9536.
