### 2026-02-02 #9536 @jjrob
@kdorheim, I have finished my review. I opened 10 issues that I believe should be addressed. Nearly all are minor and concern the documentation. I checked all of the boxes on the checklist; those check marks assume that all of the issues will be fixed (except some that are clearly just suggestions).

This is a large package with a lot of code. I ran some of the examples from several of the vignettes, and also ran the full set of `testthat` tests in the `tests` directory, as well as the full set of performance tests. Running a comprehensive test of everything is beyond what I'm able to do.

My review of the code itself was very cursory. With such a large package, I had to focus my time on checking whether it actually installed and worked. The code that I did look at looked fine. Similarly, I reviewed some of the API docs but my review was very far from comprehensive.

Finally, although I have solid expertise in general geospatial analysis, including in relevant R packages such as terra and sf, I am a marine ecologist not a terrestrial ecologist. So I cannot vouch for the correctness of implementation in any of the large number of terrestrial algorithms and analyses provided here.

Overall I think this package is a solid contribution of domain-specific code that can be used to automate tedious or frequently encountered terrestrial ecology problems and workflows. It necessarily involves a lot of functionality specific to a niche user community, such as vegetation index calculations, in addition to more general purpose functions such as the "universal spatial join". This is all to be expected and worthy of publication.
