### 2026-02-05 cwru-sdle/geospatialsuite#12 @Exelegch
Thanks... It wasn't pushed with the last correction. It is now included.
