### 2026-02-05 cwru-sdle/geospatialsuite#9 @Exelegch
Thank you for this excellent suggestion. I've revised the Summary section to be more accessible to non-specialist readers by adding an introductory sentence that provides context before diving into technical details.
