### 2026-02-05 cwru-sdle/geospatialsuite#8 @Exelegch
Thank you! Yes, the authors are correctly listed in the paper.md YAML metadata in the proper JOSS format, including all required fields (name, ORCID, affiliation, etc.). I checked JOSS repository for published articles and it seems i have done all things correctly except there must be no self generated pdf. I have removed that pdf... The final pdf by JOSS will have all parts fully integrated.
