### 2026-02-05 f5940e0
fixed doi issues and opening statement clearer in paper.md, added code of conduct, and made README better and consistent
