+++ paper/paper.bib
@@ -4,7 +4,8 @@
   journal={NASA special publication},
   volume={351},
   pages={309},
-  year={1974}
+  year={1974},
+  url={https://ntrs.nasa.gov/citations/19740022614}
 }
 
 @article{huete2002overview,
@@ -107,8 +108,10 @@
   pages={1--1},
   year={2024},
   organization={IEEE Computer Society},
-  doi={10.1109/BigData62323.2024.10825456}
+  doi={10.1109/BigData62323.2024.10825905}
 }
+
+
 
 @article{akanbi2024multiscale,
   title={Integrating multiscale geospatial analysis for monitoring crop growth, nutrient distribution, and hydrological dynamics in large-scale agricultural systems},
@@ -126,7 +129,8 @@
   title={Leveraging multimodal data for geospatiotemporal analytics},
   author={Akanbi, Olatunde David},
   year={2024},
-  school={Case Western Reserve University}
+  school={Case Western Reserve University},
+  url = {http://rave.ohiolink.edu/etdc/view?acc_num=case1699536369560882}
 }
 
 @article{xu2006modification,
