### 2026-02-09 #9536 @kdorheim
@jjrob thank you for the review! Please take a look at @Exelegch's comments & revisions to the package to determine whether they have addressed your comments and concerns.

Thank you for your comment about the terrestrial ecology element of the package needed to be reviewed by a domain-specific expert. I am searching for a third reviewer to focus on that as per your suggestion and hope to have an update soon.
