### 2026-02-09 cwru-sdle/geospatialsuite#9 @jjrob
Looks fixed by f5940e0
