### 2026-02-09 cwru-sdle/geospatialsuite#13 @jjrob
Confirmed DESCRIPTION now lists verison 0.2.0. However, @Exelegch , I recommend you also update NEWS.md:

1. Create an 0.2.0 section
2. Move the stuff from 0.1.1 that is actually in 0.2.0 to the 0.2.0 section
