### 2026-02-09 cwru-sdle/geospatialsuite#15 @jjrob
Confirmed that `vignette("performance_benchmark")` now exists in the README
