### 2026-02-09 cwru-sdle/geospatialsuite#13 @Exelegch
Thanks @jjrob This is fixed as well...

https://github.com/cwru-sdle/geospatialsuite/commit/898e4ccc509a2bfb99e6ebde17945724eb245af4
