### 2026-02-09 #9536 @Exelegch
> [@kdorheim](https://github.com/kdorheim) , you're welcome. All of the issues I opened have been fixed (and I closed them) except one (you can see status above) needs a trivial documentation update. I left that open but I'm sure [@Exelegch](https://github.com/Exelegch) can address it quickly when time permits.
>
> I did not mean to suggest that you must retain another reviewer to validate the terrestrial algorithms. I do not know if JOSS policy is to confirm correctness of implementation for things like this. There are a large number of algorithms and it could be quite time consuming to validate them exhaustively. This package looks like a collection of functions built up over a number of years that is now being published. Going back and checking everything could require a similar amount of effort. Perhaps having a terrestrial ecologist look it over and try a few basic operations would be a balanced request to make of a volunteer reviewer...

Thanks @jjrob That has been fixed as well.
