### 2026-02-09 #9536 @kdorheim
@jjrob thank you for the clarification — I agree that exhaustively validating the correctness of published algorithms is beyond the scope of this review, particularly given the number of functions and the history of the package’s development.

My suggestion was not intended to imply a requirement for full algorithmic validation. Rather, since both you and the other reviewer noted that it could be helpful to have input from a terrestrial ecologist/domain expert, I wanted to follow up on that point. A light-touch review by someone familiar with the domain — for example, checking assumptions, terminology, and a few representative use cases — seems like a reasonable and balanced approach, without implying comprehensive verification of the underlying methods.

Please let me know if that interpretation aligns with your understanding.
