### 2026-03-12 #9536 @kdorheim
@DhruvaKathuria  our review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.

The reviewer guidelines are available here: https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html. Any questions/concerns please let me know.
