### 2026-03-23 #9536 @DhruvaKathuria
@Exelegch The overall package seems really promising as it brings together a lot of functionalities. I mainly reviewed  the vegetation part of the package especially the 60 something vegetation indices which forms the core of many vegetation related functionalities. I have opened two issues: [one](https://github.com/cwru-sdle/geospatialsuite/issues/16) regarding the indices and the [other](https://github.com/cwru-sdle/geospatialsuite/issues/17) regarding implementation of the crop analysis.
