### 2026-03-23 cwru-sdle/geospatialsuite#16 @DhruvaKathuria
JOSS review [code] [documentation] Incorrect implementation of some of the vegetation indices
I found some errors with the indices which needs to be addressed. These are as follows:

-The PRI index seems to be incorrect since it is an index that uses 531 nm (green) and 570 nm (close to the green-red edge) and not the green and NIR bands as implemented in the package.
-NPQI: Again the index is incorrect because NPQI uses wavelengths 415 nm and 435 nm, which cannot be substituted with red and blue bands as implemented in the package.
-The formula for S2REP, if expanded out will always give a constant value of 722.5 irrespective of the band values
-There are also multiple instances in the indices when some of the bands are required in the formula, but they are not mentioned in the "required bands" column in the "detailed indices". Some of the examples are: PSRI, TVI, LAI. etc. This is not an exhaustive list, and there are other indices with same issue. I request you to double check the required bands in all the indices.
