### 2026-03-23 cwru-sdle/geospatialsuite#17 @DhruvaKathuria
JOSS review [code] Example to run for crop analysis
I wanted to run the crop analysis (analyze_crop_vegetation()) part of the code, but could not find an example to run and check the implementation. Is it possible to provide an example so I can quickly check the code. It is always helpful to an end user if they have a basic example to run the codes.
