### 2026-03-23 #9536 @kdorheim
@DhruvaKathuria thank you! linking those issues here for tracking purposes

* https://github.com/cwru-sdle/geospatialsuite/issues/16
* https://github.com/cwru-sdle/geospatialsuite/issues/17
