### 2026-03-25 cwru-sdle/geospatialsuite#16 @Exelegch
Thank you @DhruvaKathuria for pointing this... PRI and NPQI are narrowband hyperspectral indices — PRI requires reflectance at 531 nm and 570 nm, and NPQI at 415 nm and 435 nm. Neither can be meaningfully approximated using broadband satellite sensors (Landsat, Sentinel-2, MODIS). Both indices now issue an informative warning and return NA when called without hyperspectral narrowband input, and are clearly documented as requiring hyperspectral data.
S2REP had a mathematical error — the formula collapsed to a constant 722.5 regardless of input because the same band was used in both red-edge positions. The correct Frampton et al. (2013) formula requires two distinct red-edge bands (Sentinel-2 B5 at ~705 nm and B6 at ~740 nm). A new red_edge2 parameter has been added and the formula corrected.
Required bands were audited across all indices. All have been corrected.

https://github.com/cwru-sdle/geospatialsuite/commit/369ca61a57d7f3f31a2258c9b910f716a80f4842
