### 2026-03-25 cwru-sdle/geospatialsuite#17 @Exelegch
Thank you @DhruvaKathuria This is actually in the viginettes with no data...I have added the input data and added that to the read READme as well...

https://github.com/cwru-sdle/geospatialsuite/commit/369ca61a57d7f3f31a2258c9b910f716a80f4842
