### 2026-03-26 #9536 @DhruvaKathuria
> Hello [@DhruvaKathuria](https://github.com/DhruvaKathuria) when you have time could you please review the code/response to your github issues?

I have closed one of the issues. I will run the crop related codes tomorrow and close the issue if I find the runs satisfactory.
