### 2026-03-31 cwru-sdle/geospatialsuite#17 @DhruvaKathuria
Thanks for adding the input data. I ran the codes and the vegetation indices plot well. For the stress, yield, crop stage indices (which are mostly based on NDVI) can you add citations for how you are doing these classifications based on the NDVI thresholds.
