### 2026-03-31 cwru-sdle/geospatialsuite#17 @Exelegch
Where do want me to add the citations?

Source code or manuscript or Readme?

On Tue, Mar 31, 2026, 1:34 PM Dhruva Kathuria ***@***.***>
wrote:

> *DhruvaKathuria* left a comment (cwru-sdle/geospatialsuite#17)
> <https://github.com/cwru-sdle/geospatialsuite/issues/17#issuecomment-4164271519>
>
> Thanks for adding the input data. I ran the codes and the vegetation
> indices plot well. For the stress, yield, crop stage indices (which are
> mostly based on NDVI) can you add citations for how you are doing these
> classifications based on the NDVI thresholds.
>
