### 2026-03-31 cwru-sdle/geospatialsuite#17 @DhruvaKathuria
Anywhere that is helpful. I would say add them in the Readme and the vignette. It will be helpful for end users.
