### 2026-04-02 #9536 @DhruvaKathuria
> > Hello [@DhruvaKathuria](https://github.com/DhruvaKathuria) when you have time could you please review the code/response to your github issues?
>
> I have closed one of the issues. I will run the crop related codes tomorrow and close the issue if I find the runs satisfactory.

Thank you @Exelegch for addressing my reviews. @kdorheim I have reviewed the vegetation aspect of the package, mainly the indices and the associated crop metrics. I am satisfied with the code. It ran successfully in my test runs.
