### 2026-04-02 #9536 @Exelegch
Thank you @DhruvaKathuria

> > > Hello [@DhruvaKathuria](https://github.com/DhruvaKathuria) when you have time could you please review the code/response to your github issues?
> >
> >
> > I have closed one of the issues. I will run the crop related codes tomorrow and close the issue if I find the runs satisfactory.
>
> Thank you [@Exelegch](https://github.com/Exelegch) for addressing my reviews. [@kdorheim](https://github.com/kdorheim) I have reviewed the vegetation aspect of the package, mainly the indices and the associated crop metrics. I am satisfied with the code. It ran successfully in my test runs.

Thank you @DhruvaKathuria and other reviewers that have made this work better.
