### 2026-04-03 cwru-sdle/geospatialsuite#18 @kdorheim
DOIs for JOSS paper
@Exelegch, as part of the post-review checklist, could you please update your Bib file to include the following DOIs?  Related to https://github.com/openjournals/joss-reviews/issues/9536#issuecomment-3602346429

* terra - [10.32614/CRAN.package.terra](https://doi.org/10.32614/CRAN.package.terra)
* ggplot2 - [10.32614/CRAN.package.ggplot2](https://doi.org/10.32614/CRAN.package.ggplot2)
* leveraging multimodal data - [10.13140/RG.2.2.27132.72327](https://doi.org/10.13140/RG.2.2.27132.72327)

The editorial bot also flagged a few other citations as missing dois or no dois listed. Could you please confirm that no dois are available? If they are avaiable please add them as well.

```
 🟡 SKIP DOIs

- No DOI given, and none found for title: Monitoring vegetation systems in the Great Plains ...
- No DOI given, and none found for title: terra: Spatial Data Analysis
- No DOI given, and none found for title: ggplot2: Elegant Graphics for Data Analysis
- No DOI given, and none found for title: Cropland Data Layer
- No DOI given, and none found for title: Leveraging multimodal data for geospatiotemporal a...

❌ MISSING DOIs

- 10.1002/9781119115151.ch1 may be a valid DOI for title: Statistics for Spatial Data
```
