### 2026-04-03 cwru-sdle/geospatialsuite#18 @Exelegch
Hi @kdorheim ,

These have been fixed... Two of the references ... Monitoring vegetation systems in the Great Plains ., and Cropland Data Layer have not DOIs

https://github.com/cwru-sdle/geospatialsuite/commit/e765677bbf22e5db2384a1777c81fe0739c2e1f8
