### 2026-04-03 #9536 @Exelegch
@kdorheim ,

Double check authors and affiliations (including ORCIDs)- This is checked and fixed.
Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.--- Release made - v0.2.0
Archive the release on Zenodo/figshare/etc and post the DOI here. -Done (https://doi.org/10.5281/zenodo.19408273)
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.- This is correct as well
Make sure that the license listed for the archive is the same as the software license.- Checked
