### 2026-04-05 #9536 @Exelegch
@kdorheim It seems the bot isn't checking the updated paper. Missing DOIs still persistent.
