### 2026-04-05 #9536 @kdorheim
@Exelegch it looks like you have two copies of paper.md and paper.bib one in your docs and another in the paper directory. Could you remove the files from the directory you have not been updating?
