### 2026-04-06 cwru-sdle/geospatialsuite#19 @kdorheim
Final JOSS Manuscript Edits
Related to https://github.com/openjournals/joss-reviews/issues/9536#issuecomment-4183882997

Please address the following

* L24 in the compiled manuscript there is a typo "functionalityâ€”" please fix
