### 2026-04-06 #9536 @kdorheim
@Exelegch please address https://github.com/cwru-sdle/geospatialsuite/issues/19 for final manuscript edits
