### 2026-04-06 #9536 @Exelegch
> [@Exelegch](https://github.com/Exelegch) please address [cwru-sdle/geospatialsuite#19](https://github.com/cwru-sdle/geospatialsuite/issues/19) for final manuscript edits

Thank you @kdorheim It has been fixed
