### 2026-04-06 cwru-sdle/geospatialsuite#19 @kdorheim
@Exelegch In L24 it still looks like there is a `”` that is either unneeded or is missing the first set of `"`
