### 2026-04-06 cwru-sdle/geospatialsuite#19 @Exelegch
Hi @kdorheim I saw that after my push and reveiw but I have fixed that before you noticed... Sorry for the typo.
